refactor: split analytics into schema, redaction, validation and environment
