+++ src/main/services/analytics.js
@@ -1,1227 +1,39 @@
 // the only place telemetry leaves the app. the renderer routes through ipc to
-// here, so the opt-out gate, the allowlist and the redaction below cannot be
+// here, so the opt-out gate, the allowlist and the redaction cannot be
 // bypassed.
+//
+// the three of those live beside this file rather than in it: analytics/
+// schema.js is what each event may carry and in what vocabulary, analytics/
+// redact.js the sanitizer for the one free-text property, analytics/validate.js
+// the kind check every value passes through, and analytics/environment.js the
+// facts about this install. this file is the class that consults them, and the
+// only export surface any of it has.
 
-const os = require("os")
-const { APP_CONFIG, SUPPORTED_PLATFORMS } = require("../utils/constants")
+const { APP_CONFIG } = require("../utils/constants")
 const { redactLogLine } = require("./ytdlp-engine")
-// the ceiling on any one playlist link, which is what bounds every item count
-// below. imported rather than written out: a cap that moves has to move the
-// validation with it, or every count above the old one is dropped in silence
-const { PLAYLIST_MAX_ITEMS } = require("../utils/ytdlp-mappers")
 const { describeError, getAppVersion } = require("../utils/analytics-helpers")
 const {
-  ERROR_CATEGORIES,
-  ERROR_STAGES
-} = require("../utils/error-taxonomy")
-
-/**
- * every event we send, and every property it may carry. anything absent is
- * dropped rather than forwarded.
- *
- * this is the privacy contract, not documentation of it. a denylist would
- * only stop the leaks we thought of; this stops the ones a later caller
- * invents - a url, a title, a filename, a path - because it has to be added
- * here on purpose before it can leave. the renderer's events arrive over ipc
- * as an untrusted property bag, and this is what makes forwarding them safe.
- *
- * adding an event or a property here is a deliberate privacy decision. if a
- * later task finds this list in its way, that is the point of it.
- */
-const ALLOWED_PROPERTIES = {
-  app_launched: [
-    "is_first_launch",
-    "previous_version",
-    "engine_version",
-    // the standing figure: what share of installs are signed in right now,
-    // as opposed to cookies_imported's "who ever imported"
-    "cookies_signed_in"
-  ],
-  url_submitted: ["platform", "url_kind"],
-  media_info_loaded: [
-    "platform",
-    "duration_bucket",
-    "formats_count",
-    "load_ms_bucket",
-    "used_cookies",
-    // a playlist listing instead of a video's: it has no duration and no format
-    // list, and what it has instead is a length. bucketed, never raw
-    "playlist_size"
-  ],
-  media_info_failed: [
-    "platform",
-    "error_category",
-    "error_stage",
-    "error_message",
-    "used_cookies"
-  ],
-  /**
-   * which half of a `watch?v=…&list=…` link the user meant.
-   *
-   * the one event this taxonomy owes to url_kind's "playlist" value, which
-   * counts how often such a link is pasted and cannot say what was wanted. two
-   * properties and no more: the answer, and how big the list they were choosing
-   * about was.
-   */
-  playlist_prompt_answered: ["choice", "playlist_size"],
-  /**
-   * somebody took the helper line up on its offer of a playlist.
-   *
-   * the platform and nothing else, and the platform is a constant here: the
-   * link is ours rather than one the user pasted, so there is nothing about it
-   * to report. what it answers is whether telling people playlists exist is
-   * what makes them try one, which is a count of clicks against the
-   * url_kind: "playlist" submissions that follow.
-   */
-  playlist_hint_clicked: ["platform"],
-  download_started: [
-    "platform",
-    "media_type",
-    "quality",
-    "is_trimmed",
-    "audio_format",
-    /**
-     * a playlist is one download of n videos, so it extends this event rather
-     * than sending one beside it: the same platform, media type and requested
-     * quality, plus that it is a playlist and how many videos were ticked.
-     *
-     * absent rather than false on a single video. every existing event keeps
-     * exactly the properties it has always had, which is what lets the two be
-     * compared at all - and "not a playlist" is what no is_playlist means.
-     */
-    "is_playlist",
-    "item_count"
-  ],
-  // download_started's properties plus the completion measures, which is what
-  // makes the two ends of the funnel join on the same dimensions
-  download_completed: [
-    "platform",
-    "media_type",
-    "quality",
-    "is_trimmed",
-    "audio_format",
-    "file_size_mb",
-    "elapsed_bucket",
-    "speed_bucket",
-    "used_cookies",
-    /**
-     * what the playlist run actually did, and this is the event that can say.
-     *
-     * a run that saved eight of nine videos exits 1 and is a completion all the
-     * same, so a non-zero `items_skipped` here is the normal shape of a partial
-     * success rather than a failure in disguise. `items_reused` is its own
-     * number and is never folded into `items_saved`: an archive skip records
-     * that some earlier run wrote the file, not that this one did.
-     */
-    "is_playlist",
-    "items_saved",
-    "items_reused",
-    "items_skipped",
-    "items_total"
-  ],
-  // likewise download_started's properties plus its own. a schema consistent
-  // on success and silent on failure would be worse than either answer applied
-  // to both ends
-  download_failed: [
-    "platform",
-    "media_type",
-    "quality",
-    "is_trimmed",
-    "audio_format",
-    "error_category",
-    "error_stage",
-    "error_message",
-    "progress_at_failure",
-    "used_cookies",
-    /**
-     * what a failed playlist had already written, which is the half a user can
-     * still see on disk.
-     *
-     * two counts and not four, deliberately. a run that broke halfway never
-     * reached the videos behind the break, so calling them skipped would be a
-     * guess dressed as a measurement - and a reuse count is about an archive
-     * that a run this broken may never have read.
-     */
-    "is_playlist",
-    "items_saved",
-    "items_total"
-  ],
-  // the same two counts, for the same reason: a cancel is a kill, and the
-  // videos it had already finished are still there
-  download_cancelled: [
-    "platform",
-    "media_type",
-    "progress_at_cancel",
-    "is_playlist",
-    "items_saved",
-    "items_total"
-  ],
-  engine_seeded: ["reason", "engine_version", "elapsed_bucket"],
-  engine_updated: ["from_version", "to_version"],
-  engine_update_failed: ["update_reason", "error_message"],
-  cookies_imported: ["success", "has_youtube_cookies", "signed_in"],
-  // the two halves of the coffee prompt: how many were shown one, and how many
-  // acted on it. Both carry the milestone, so the answer can be read per step
-  // rather than only in total - which is what says whether the later ones are
-  // worth keeping
-  support_prompt_shown: ["milestone"],
-  support_prompt_clicked: ["milestone"]
-}
-
-// built once, so a capture is a set lookup rather than a scan
-const ALLOWED_BY_EVENT = new Map(
-  Object.entries(ALLOWED_PROPERTIES).map(([event, keys]) => [
-    event,
-    new Set(keys)
-  ])
-)
-
-/**
- * the shape each property is allowed to have. an allowed name is only half a
- * privacy boundary - `platform` set to a url would have left verbatim.
- *
- * see the note above the patterns below for why the string kinds are
- * vocabularies and numeric grammars rather than one general-purpose charset.
- */
-const PROPERTY_KINDS = {
-  // flags
-  is_first_launch: "bool",
-  is_trimmed: "bool",
-  is_playlist: "bool",
-  success: "bool",
-  has_youtube_cookies: "bool",
-  signed_in: "bool",
-  cookies_signed_in: "bool",
-  // youtube only, and on both ends of a download plus the lookup - the join
-  // that answers whether cookies actually moved the refusal rate
-  used_cookies: "bool",
-
-  // counts and measures
-  formats_count: "number",
-  file_size_mb: "number",
-  progress_at_failure: "number",
-  progress_at_cancel: "number",
-  // which step of the sequence a prompt was, so 5 and 100 stay distinguishable
-  milestone: "number",
-
-  /**
-   * a playlist run's own arithmetic, and all five are raw rather than bucketed.
-   *
-   * they describe the operation rather than the person: "8 saved, 2 already
-   * downloaded, 1 skipped of 11" is the outcome, and rounding it into buckets
-   * would destroy exactly the arithmetic these exist for. how big the playlist
-   * they came from was is a different question, and that one is NOT raw - see
-   * playlist_size below.
-   *
-   * their own kind rather than "number", because the bound is real and knowable:
-   * a run covers at most PLAYLIST_MAX_ITEMS videos, so a whole number from 0 to
-   * the cap is the whole range, where the generic kind would take a fractional
-   * 1e9. narrowing a kind to what the feature guarantees is the cheapest
-   * validation there is, and the reviewer that asked for it is right that
-   * "small integer" is not a claim a 0-to-a-billion check makes.
-   */
-  item_count: "count",
-  items_saved: "count",
-  items_reused: "count",
-  items_skipped: "count",
-  items_total: "count",
-
-  // a controlled vocabulary that normalizes instead of dropping
-  platform: "platform",
-
-  // finite vocabularies, listed in PROPERTY_VOCABULARIES
-  url_kind: "vocabulary",
-  media_type: "vocabulary",
-  audio_format: "vocabulary",
-  choice: "vocabulary",
-
-  /**
-   * how many videos the pasted playlist holds, as one of four labels.
-   *
-   * a **vocabulary** and not a bucket, which was a correction: the generic
-   * bucket grammar is "a digit or a comparison and a short unit", and it
-   * forwarded "1984 film", a video-id-shaped "123456789abC" and the exact
-   * length "5283 vids" untouched. that the renderer's own helper only produces
-   * safe labels is not an argument this module accepts - the renderer is the
-   * least-trusted caller here by construction, and its bag arrives over ipc.
-   *
-   * the values are ours and there are four of them, which is precisely the case
-   * the note below says a vocabulary is for. bucketing it at all is still the
-   * right call for the value itself: the exact length of a list, next to a
-   * platform and a locale, is close to naming which list it was.
-   */
-  playlist_size: "vocabulary",
-  reason: "vocabulary",
-  update_reason: "vocabulary",
-  error_category: "vocabulary",
-  error_stage: "vocabulary",
-
-  // part grammar, part vocabulary - see checkKind
-  quality: "quality",
-
-  // version strings, which must lead with a digit - that is what a filename
-  // cannot do
-  previous_version: "version",
-  engine_version: "version",
-  from_version: "version",
-  to_version: "version",
-
-  // pre-bucketed measurements, never a raw one
-  duration_bucket: "bucket",
-  elapsed_bucket: "bucket",
-  speed_bucket: "bucket",
-  load_ms_bucket: "bucket",
-
-  // free text, scrubbed and clipped
-  error_message: "text",
-
-  /**
-   * the super properties.
-   *
-   * these are never named in ALLOWED_PROPERTIES, because no caller may send
-   * one - init() builds them and capture() spreads them last, over the
-   * validated bag. they are kinded all the same: that placement is exactly
-   * what makes them worth checking, since one of them rides every event this
-   * app will ever send, and a leak there is the most amplified one available.
-   */
-  app_version: "version",
-  os: "vocabulary",
-  os_version: "version",
-  arch: "vocabulary",
-  locale: "locale",
-
-  /**
-   * whether this install can mint a PO token at all, which is two separate
-   * questions: is there a js runtime to run the generator on, and is the
-   * payload actually installed. neither is answerable from a single event -
-   * they describe the machine, not the operation - and both have to ride every
-   * event to be joinable against the failures they explain.
-   *
-   * media_info_failed is the one that matters: a renderer event, raised by the
-   * half of the app that cannot see either of these facts.
-   */
-  deno_present: "bool",
-  pot_provider: "bool",
-
-  // two values, both ours, listed in PROPERTY_VOCABULARIES - a closed kind,
-  // which is what a build indicator should be: it exists so a dev session is
-  // filterable rather than indistinguishable, and nothing about that needs a
-  // kind that admits anything else
-  environment: "vocabulary"
-}
-
-const KIND_BY_PROPERTY = new Map(Object.entries(PROPERTY_KINDS))
-
-/**
- * there is no general-purpose grammar for a short label, and six rounds of
- * counterexamples were the proof. a charset wide enough for "1-5 min" admits
- * "My Holiday Video"; narrow it to a lowercase identifier and it still admits
- * "holiday". a single word is indistinguishable from a title by inspection,
- * because it may *be* one.
- *
- * so every string property is now one of three things, and none of them is a
- * general grammar:
- *
- *   - a vocabulary, when the values are ours and finite
- *   - a numerically anchored grammar, when the values genuinely grow but only
- *     along a numeric axis (heights, bitrates, versions, buckets)
- *   - redacted text, when the value is free prose and treated as such
- *
- * the enumeration cost is real, and it is the same cost ALLOWED_PROPERTIES
- * already pays on property names: a later task that sends something new has
- * to come here and say so. that friction is the feature.
- */
-
-// heights and bitrates grow with the format list, so they are matched rather
-// than listed - but both are anchored on digits, which "holiday" cannot fake
-const QUALITY_HEIGHT_PATTERN = /^\d{2,4}p$/
-const QUALITY_BITRATE_PATTERN = /^\d{1,4}kbps$/
-
-// everything else extractQuality() can return (analytics-helpers.js). these
-// are fixed outputs of its own branches, not values a format list supplies,
-// so they are listed rather than matched
-const QUALITY_VALUES = new Set([
-  "audio",
-  "best_audio",
-  "best_available",
-  "high_quality",
-  "low_quality",
-  "m4a",
-  "medium_quality",
-  "mp3",
-  "original_audio",
-  "unknown"
-])
-
-// digits and dots, with an optional prerelease tail. leading with a digit was
-// not enough on its own: "2024.mp4" leads with one too.
-//
-// letters are confined to the prerelease's *first* segment; every later
-// dot-segment is numeric. an earlier version of this allowed dots inside the
-// tail, which let "1-holiday.mp4" through - "1" satisfied the head and the
-// extension rode along behind the dash.
-const VERSION_PATTERN =
-  /^[0-9][0-9.]{0,30}(?:-[A-Za-z0-9]{1,16}(?:\.[0-9]+)*)?$/
-
-// a pre-bucketed measurement: leads with a digit or a comparison, ends in a
-// short unit. "1-5 min", "<1m", "10-50 MB", "60+ min"
-//
-// the trailing `+` is admitted because an open-ended top bucket is written
-// that way more often than any other, and a grammar that forbids the reflex
-// spelling buys nothing: ">60 min" carries the identical information, so the
-// only thing rejecting "60+ min" achieves is a dropped event behind a warning
-// production never shows anyone. still anchored on a digit or a comparison,
-// so no title or filename becomes reachable.
-//
-// the unit is required, not optional: without it a bare "2024" was a valid
-// bucket, and a unitless label is unreadable in a chart anyway - so demanding
-// one tightens the boundary and improves the data at the same time.
-const BUCKET_PATTERN = /^[<>]?[0-9]+(-[0-9]+)?\+?\s?[a-zA-Z]{1,4}$/
-
-/**
- * a bcp-47 language tag, anchored on case and length.
- *
- * this one is neither a vocabulary nor numerically anchored, and the reason is
- * that the values are chromium's rather than ours: app.getLocale() returns
- * whichever ui locale it resolved, and that set grows with electron upgrades.
- * a listed set would silently stop reporting the locale of everybody who ran
- * the version that added one - a drop nobody triggered and nobody would see,
- * which is a different thing from the deliberate friction the other lists buy.
- *
- * so it is anchored on shape, but on a shape prose does not have: two or three
- * lowercase letters, then at most a titlecase four-letter script and a
- * two-letter uppercase or three-digit region. "my-video" fails on the second
- * segment's case and "holiday" on the first one's length, which is the pair
- * that defeated every general-purpose slug grammar tried before this.
- */
-const LOCALE_PATTERN =
-  /^[a-z]{2,3}(?:-[A-Z][a-z]{3})?(?:-(?:[A-Z]{2}|[0-9]{3}))?$/
-
-// what readLocale() reports when electron is not there to ask
-const LOCALE_UNKNOWN = "unknown"
-
-// the two builds that exist: one somebody installed, and one run from source
-// with the dev opt-in set. see the environment vocabulary below
-const ENVIRONMENT_PRODUCTION = "production"
-const ENVIRONMENT_DEVELOPMENT = "development"
-
-/**
- * the finite vocabularies, one per property.
- *
- * the two error vocabularies are read from the taxonomy module rather than
- * copied, so a category added there is accepted here for free. the rest are
- * listed because their sources are scattered literals rather than an exported
- * table - each is noted with where it comes from, so it can be re-derived.
- *
- * an empty set is a deliberate state, not an oversight: it means the values
- * do not exist yet, and the first task to send one has to come here first.
- */
-const PROPERTY_VOCABULARIES = {
-  error_category: new Set(Object.values(ERROR_CATEGORIES)),
-  error_stage: new Set(Object.values(ERROR_STAGES)),
-
-  // node's whole documented range for process.platform and process.arch. both
-  // are genuinely finite and genuinely not ours to invent, and node adds to
-  // them about once a decade - which is the case a list is for. the app runs
-  // on three of these; the rest are here because this set claims to be the
-  // whole list, and a drop on an unusual platform would take that install's
-  // os off every event it ever sends
-  os: new Set([
-    "aix",
-    "android",
-    "cygwin",
-    "darwin",
-    "freebsd",
-    "haiku",
-    "linux",
-    "netbsd",
-    "openbsd",
-    "sunos",
-    "win32"
-  ]),
-  arch: new Set([
-    "arm",
-    "arm64",
-    "ia32",
-    "loong64",
-    "mips",
-    "mipsel",
-    "ppc",
-    "ppc64",
-    "riscv64",
-    "s390",
-    "s390x",
-    "x64"
-  ]),
-
-  // the runner emits "combined" for a video download - an ffmpeg detail
-  // meaning two streams were merged - and task 6 normalizes that to "video"
-  // before sending. the raw term is deliberately absent rather than missing:
-  // this taxonomy answers whether people take video or audio, and should not
-  // carry a word that describes how the engine assembled the file
-  media_type: new Set(["video", "audio"]),
-
-  // the keys of AUDIO_MODE_PRESETS - normalizeAudioMode returns the key, not
-  // the codec it maps to (both in ytdlp-engine.js)
-  audio_format: new Set(["mp3", "m4a", "original"]),
-
-  // every reason ytdlp-updater can report: its `reason:` literals, the three
-  // passed positionally to installDirectory(), and "download-failed", which is
-  // written as a ternary arm in downloadAndUnpack (ytdlp-updater.js) and so is
-  // invisible to a grep for the keyword. engine_seeded only ever sends the
-  // three positional ones - the rest are here because this set claims to be
-  // the whole list
-  reason: new Set([
-    "asset-layout-unexpected",
-    "bundled-newer",
-    "busy",
-    "cancelled",
-    "check-failed",
-    "checksum-mismatch",
-    "checksum-missing",
-    "completed",
-    "copy-failed",
-    "corrupt",
-    "corrupt-and-no-bundle",
-    "download-failed",
-    "engine-dir-failed",
-    "missing",
-    "no-binary-available",
-    "probe-failed",
-    "repaired",
-    "swap-failed",
-    "swap-stranded",
-    "unsupported-platform",
-    "up-to-date",
-    "version-mismatch"
-  ]),
-
-  // why an engine update did not happen: every reason checkForUpdate() can
-  // return except the two that mean nothing went wrong, "up-to-date" and
-  // "completed", which index.js sends no event for at all. read off its return
-  // sites rather than seed()'s - the seeding-only reasons (missing, corrupt,
-  // bundled-newer, copy-failed, engine-dir-failed, corrupt-and-no-bundle)
-  // cannot reach this property, because runUpdateLocked ignores what
-  // seedLocked said and returns its own.
-  //
-  // "busy" is one of these on purpose: the check refused because a download
-  // held the engine gate, so it never ran - and an install whose engine
-  // therefore never updates is exactly what this event exists to find
-  update_reason: new Set([
-    "asset-layout-unexpected",
-    "busy",
-    "cancelled",
-    "check-failed",
-    "checksum-mismatch",
-    "checksum-missing",
-    "download-failed",
-    "no-binary-available",
-    "probe-failed",
-    "repaired",
-    "swap-failed",
-    "swap-stranded",
-    "unsupported-platform",
-    "version-mismatch",
-
-    // the one value the updater does not produce. checkForUpdate() can reject
-    // rather than return - runUpdateLocked leaves probeVersion and the
-    // recovery renames unguarded - and index.js names that itself. kept apart
-    // from "check-failed", which is the tag lookup failing and returning
-    // normally: merging them would hide a bug in our own code inside the
-    // commonest network blip there is
-    "check-rejected"
-  ]),
-
-  /**
-   * which half of an ambiguous link the user meant.
-   *
-   * the two values of MixedLinkChoice (renderer/src/lib/mixedLinkStore.ts), and
-   * there cannot be a third: the question has two buttons, and closing it
-   * without pressing either sends nothing at all - an abandoned paste teaches
-   * us nothing and is not an answer.
-   *
-   * named `choice` rather than `playlist_choice` because the event already says
-   * what was being chosen, and it is a closed vocabulary either way: nothing a
-   * later caller invents can reach this property without being listed here.
-   */
-  choice: new Set(["video", "playlist"]),
-
-  /**
-   * how big the pasted playlist is, as one of four labels.
-   *
-   * these are PLAYLIST_SIZE_BUCKET_LABELS in renderer/src/lib/analytics.ts,
-   * written out here rather than derived because there is no way to import them:
-   * they live in the renderer's own build. so this is the mirror, and it is the
-   * authority - a label the renderer adds without adding it here is dropped, and
-   * that is the direction this list is meant to fail in.
-   *
-   * the two boundaries that are load-bearing are the item cap (100) and the fact
-   * that the top bucket is open: a channel of 5,283 videos and a channel of
-   * 300 are the same answer to "did somebody paste something enormous", and one
-   * exact figure would say which channel.
-   */
-  playlist_size: new Set([
-    "1-5 vids",
-    "6-25 vids",
-    "26-100 vids",
-    ">100 vids"
-  ]),
-
-  // the shape of a submitted link, named by urlKind() in the renderer's
-  // lib/analytics.ts - these are ours to invent, and this is the whole list.
-  //
-  // "playlist" is the one that pays for the rest: we take the single video out
-  // of a playlist url, and how often somebody expected otherwise is a question
-  // nothing else here answers. "short-link" is the redirect hosts (youtu.be,
-  // pin.it, vm/vt.tiktok.com, tiktok.com/t), which cost a resolution step
-  // before anything else can happen
-  url_kind: new Set(["video", "shorts", "playlist", "short-link", "embed"]),
-
-  /**
-   * which kind of build sent the event.
-   *
-   * "development" is a session somebody opted into with CLIPLY_ANALYTICS_DEV=1
-   * - a packaged build is always "production", whatever its environment says,
-   * because app.isPackaged is what both this and the gate above read.
-   *
-   * it is here because the gate failing shut is invisible without it: a
-   * dashboard receiving nothing from real installs looks exactly like a
-   * dashboard receiving nothing at all, and that is how a build that had sent
-   * zero events since the feature landed went unnoticed through nine rounds of
-   * review. one dimension makes the two distinguishable.
-   */
-  environment: new Set([ENVIRONMENT_PRODUCTION, ENVIRONMENT_DEVELOPMENT])
-}
-
-const VOCABULARY_BY_PROPERTY = new Map(Object.entries(PROPERTY_VOCABULARIES))
-
-// what an unrecognised platform becomes. it is not a drop - the event still
-// arrives, tagged - so the bucket counts how often people try a site we do not
-// support. it does not say which sites those are: every unrecognised value
-// collapses to this one literal, and no hostname is kept anywhere, which is
-// what the no-url rule requires. an aggregate, not a list.
-const PLATFORM_UNSUPPORTED = "unsupported"
-
-/**
- * the platforms a value may name.
- *
- * SUPPORTED_PLATFORMS is imported rather than copied, but it is not the whole
- * set: it lists youtube, instagram and tiktok, while the engine's own
- * download list (SUPPORTED_DOWNLOAD_PLATFORMS in ipc-handlers.js) lists
- * youtube, pinterest and tiktok. the two are not mirrors. pinterest is fully
- * supported - it has dedicated handling in ipc-handlers and its own
- * extractQuality mapping - so validating against SUPPORTED_PLATFORMS alone
- * would relabel a real platform as unsupported.
- *
- * pinterest is therefore the one name written out here. that list is
- * module-local to ipc-handlers and cannot be imported: analytics is about to
- * become one of ipc-handlers' dependencies, so requiring it back would be a
- * cycle, and it pulls in electron besides. the real fix is one list instead
- * of two, which is not this task's to make.
- */
-const KNOWN_PLATFORMS = new Set([
-  ...Object.keys(SUPPORTED_PLATFORMS).map((key) => key.toLowerCase()),
-  "pinterest",
-  "unknown",
-  PLATFORM_UNSUPPORTED
-])
-
-const MAX_TEXT_LENGTH = 500
-const MAX_NUMBER = 1e9
-
-// how long the background drain an opt-out starts is given before it gives up.
-// nobody waits on it, so this is not about responsiveness - it is what stops a
-// detached client sitting on a dead socket indefinitely. the same two seconds
-// the quit path allows (QUIT_FLUSH_TIMEOUT_MS in index.js), because the amount
-// of unsent telemetry worth holding a resource for is the same either way.
-const OPT_OUT_FLUSH_TIMEOUT_MS = 2000
-
-/**
- * error_message is the one property whose values cannot be enumerated.
- *
- * every other string kind is answered by a vocabulary or a numeric anchor,
- * because - as six rounds of counterexamples established - a pattern cannot
- * establish what a string *is*, only what it still looks like. this property
- * has neither defence available to it: it is free text by design, and the
- * useful half of it is precisely the wording nobody wrote for us. an error
- * that did not come from the engine carries whatever node, yt-dlp or a throw
- * site put in it, and the updater's own http failures name their urls.
- *
- * redactLogLine is not the answer either. it serves the local log file, where
- * an absolute path is legitimately useful to whoever is debugging their own
- * machine; it scrubs the home directory and a query string and nothing else.
- * merging the two thresholds would degrade the logs to protect telemetry.
- *
- * so this works the other way round from a grammar. scrub the shapes we can
- * name, then ask whether anything location-shaped or identity-shaped is still
- * standing, and if so send a fixed placeholder rather than the text. an
- * incomplete pattern list then costs a message instead of a leak, which is the
- * direction this property has to fail in.
- */
-const TEXT_PLACEHOLDER = "[redacted]"
-
-/**
- * how far a path runs once it has started.
- *
- * paths hold spaces - "My Holiday Video.mp4" is one token to a person and
- * three to a regex - and stopping at the first space is what leaves the words
- * of a title standing after the filename at the end of them has gone. so a
- * space is consumed only while something further along the line still carries
- * a separator or a dot, which keeps the path whole without swallowing the
- * sentence that follows it.
- */
-const PATH_TAIL = String.raw`(?:[^\s'"<>|,;)]|[ \t](?=[^\n\r]*[\\/.]))*`
-
-const PATH_HEADS = [
-  // a windows unc share
-  String.raw`\\\\`,
-  // a drive letter. the word boundary keeps "Note:" and a clock time out
-  String.raw`\b[A-Za-z]:[\\/]`,
-  // what redactLogLine leaves behind where a home directory was
-  String.raw`~[\\/]`,
-  // two separators, so "MiB/s" and "and/or" stay prose rather than becoming
-  // locations. a *relative* path is deliberately unreachable here - it is
-  // indistinguishable from prose, and the refusal below is what answers it
-  String.raw`/[^\s'"<>|,;)]*/`
-]
-
-/**
- * the identifier a credential's name is a part of, wherever in it that is.
- *
- * this rule has now been wrong twice for the same reason, and the reason is
- * worth stating rather than patching around a third time. it anchored on `\b`,
- * which missed `client_secret` because "_" is a word character. it then
- * anchored on a prefix, which missed `aws_secret_access_key` - a standard aws
- * field - because the marker turned out to sit in the middle. both fixes moved
- * the position the marker was assumed to occupy; neither questioned that it
- * occupied a fixed one.
- *
- * it does not. `client_secret`, `aws_secret_access_key`, `cookie_store`,
- * `x-api-key`, `signing_secret`: the marker is a *component* of a compound
- * name, and a component can be anywhere. so the rule is now the whole of that
- * claim - a marker appearing anywhere in a key that is being assigned a value -
- * and there is no position left for a later counterexample to exploit.
- *
- * the same shape as task 6's ipv6 rule, which required every group to be
- * non-empty and so missed `::` - which is how one is actually written. that
- * fix, like this one, was to stop describing where the thing sits and start
- * matching what it is.
- *
- * the punctuation in the class is the second half of the same lesson, learned
- * the same way. a flat identifier is one spelling of a key and not the only
- * one: `credentials[secret]`, `creds[0][secret]`, `config%5Bapi_key%5D`,
- * `"api_key"`, `'secret'` and `auth.client_secret` are all one key with a
- * marker in it, written by a form serializer, a query encoder or a json dump.
- * enumerating those serializations would be the position mistake again in a
- * different costume, so the class says what a key character is and lets any
- * arrangement of them count.
- *
- * the lookbehind rather than `\b` is what makes the match start at the real
- * beginning of the key. with `\b` a quoted key began at the letter and left
- * its opening quote stranded, which then had nothing to pair with.
- *
- * `[=:]` on the key is what still keeps ordinary prose out, and that half has
- * held through every version of this rule.
- */
-const CREDENTIAL_KEY_CHAR = String.raw`[A-Za-z0-9_\-.%\[\]"']`
-const CREDENTIAL_KEY_PART = CREDENTIAL_KEY_CHAR + "*"
-const CREDENTIAL_KEY_START = String.raw`(?<!${CREDENTIAL_KEY_CHAR})`
-
-/**
- * and the value, which may be quoted or bare. a json dump writes
- * `"secret": "hunter2"`, and stopping at the opening quote would publish
- * everything after it.
- */
-const CREDENTIAL_VALUE = String.raw`(?:"[^"]*"|'[^']*'|[^\s'"&]+)`
-
-const CREDENTIAL_NAMES = [
-  "api[_-]?key",
-  "access[_-]?token",
-  "refresh[_-]?token",
-  "auth[_-]?token",
-  "id[_-]?token",
-  "token",
-  "secret",
-  "password",
-  "passwd",
-  "pwd",
-  "session[_-]?id",
-  "cookie"
-].join("|")
-
-const TEXT_SCRUBS = [
-  // a scheme url first, so its own path is consumed here rather than by the
-  // path patterns below. brackets end it so a parenthesised url reads
-  // cleanly; angle brackets do not, because redactLogLine leaves a
-  // "?<redacted>" tail behind that belongs to the url it came from
-  [/\b[a-z][a-z0-9+.-]*:\/\/[^\s'"()[\]]+/gi, "[url]"],
-  [/\bwww\.[^\s'"()[\]]+/gi, "[url]"],
-
-  /**
-   * a credential.
-   *
-   * an opaque token is prose to every rule below this one - no separator, no
-   * dot, no host - which is the same position a bare title is in, and the
-   * floor this module accepts for titles. it is admitted here anyway, because
-   * the two are not the same thing twice: a token has a findable edge that a
-   * title does not. no wording of ours, of node's or of yt-dlp's says "Bearer"
-   * or "api_key=", so keying on the marker costs nothing, while a leaked
-   * credential is a different category of harm from a leaked video title.
-   *
-   * an Authorization value is a secret in its entirety, so that one runs to
-   * the end of the line rather than to the end of a token - a scheme and its
-   * payload are two tokens, and stopping between them would publish the half
-   * that matters. "bearer" and "negotiate" are also matched without their
-   * header name; "basic" and "digest" deliberately are not, because unlike
-   * those two they are ordinary english and would eat the word after them.
-   *
-   * see CREDENTIAL_KEY_PART above for why the name is matched as a component
-   * of the key rather than at any particular end of it.
-   */
-  [
-    new RegExp(
-      CREDENTIAL_KEY_START +
-        String.raw`${CREDENTIAL_KEY_PART}authorization${CREDENTIAL_KEY_PART}\s*[=:].*`,
-      "gi"
-    ),
-    "[credential]"
-  ],
-  [/\b(?:bearer|negotiate)\s+[^\s'"]+/gi, "[credential]"],
-  [
-    new RegExp(
-      CREDENTIAL_KEY_START +
-        String.raw`${CREDENTIAL_KEY_PART}(?:${CREDENTIAL_NAMES})${CREDENTIAL_KEY_PART}\s*[=:]\s*${CREDENTIAL_VALUE}`,
-      "gi"
-    ),
-    "[credential]"
-  ],
-
-  [/[^\s'"<>@]+@[^\s'"<>@]+\.[^\s'"<>@]+/g, "[email]"],
-
-  // an address. the updater says "could not connect to <address> for <url>",
-  // and while the address it means is github's, the check below cannot tell
-  // one of those from the machine's own: an ipv4 carries no letters, so the
-  // filename rule reads it as harmless and lets it through. four groups of
-  // digits is more than a version string has, so "2026.08.19" is untouched
-  [/\b\d{1,3}(?:\.\d{1,3}){3}\b/g, "[address]"],
-
-  // ipv6, matched as "hex and colons around a ::" rather than by validating
-  // the grammar. validating it is what went wrong the first time: requiring
-  // every group after a colon to be non-empty excluded ::1, 2001:db8::1 and
-  // fe80::a - which is to say every compressed form, which is to say every
-  // ipv6 anyone writes. a generous match here can only ever replace more.
-  // the brackets are the url spelling and %zone is a link-local's interface,
-  // which names the machine's own hardware
-  [/\[?[0-9a-f:]*::[0-9a-f:]*(?:%[A-Za-z0-9._-]+)?\]?/gi, "[address]"],
-  // and the uncompressed eight-group form, which has no :: to find
-  [/\b(?:[0-9a-f]{1,4}:){7}[0-9a-f]{1,4}\b/gi, "[address]"],
-
-  ...PATH_HEADS.map((head) => [new RegExp(head + PATH_TAIL, "g"), "[path]"]),
-
-  // whatever is left carrying an extension. \p{L} rather than [A-Za-z]: a
-  // title in another script is still a title
-  [/\S*\p{L}\S*\.[\p{L}\p{N}]{1,6}(?=$|[\s.,;:!?)\]'"])/gu, "[file]"]
-]
-
-// a span that holds nothing but markers has already been cleaned, and keeping
-// it is what leaves node's "open '[path]'" readable instead of "open [text]"
-const MARKERS_ONLY =
-  /^(?:\s|\[(?:url|path|file|email|address|credential|text)\])*$/
-
-/**
- * what may not survive the scrub above.
- *
- * deliberately broader than it: this half decides whether the text may be sent
- * at all, so it asks whether the string still looks like a location or an
- * identity rather than whether we know how to clean it.
- */
-const TEXT_UNSAFE = [
-  // any separator still standing is one the scrubs could not account for -
-  // a relative path, a windows path in a shape we did not parse
-  /[/\\]/,
-  /@/,
-  // a dotted token carrying a letter: a filename whose extension the scrub
-  // did not recognise. a version string is digits either side and survives
-  /\S*\p{L}\S*\.[\p{L}\p{N}]/u,
-
-  // two colons with nothing but hex between them: an address in a spelling
-  // the scrub above did not name. this is the general answer to the shape,
-  // rather than a longer list of address grammars - a mac address is not an
-  // ip and no rule above would have caught one.
-  //
-  // it costs the messages that carry a second colon of their own: a
-  // "10:30:45" clock, an ffmpeg "00:01:23" duration, an iso timestamp. one
-  // colon is untouched, so "ERROR:", "reason: value", "code 137: Killed",
-  // "10:30" and a windows drive letter all still send
-  /[0-9a-f]*:[0-9a-f]*:/i,
-  // something hanging off a home marker. a bare "~" is a whole home directory
-  // and says nothing; "~something" is a path we did not parse
-  /~\S/
-]
-
-/**
- * replace the quoted spans a message uses to name the thing it is about
- *
- * a title, an argument, a filename: whatever is still inside a span after the
- * scrubs above ran is text nobody vouched for, so the span goes as a whole.
- *
- * @param {string} text - already scrubbed of the shapes we can name
- * @returns {string} the same text with unvouched-for spans replaced
- */
-function scrubQuotedSpans(text) {
-  return (
-    text
-      .replace(/"([^"]*)"/g, (span, inner) =>
-        MARKERS_ONLY.test(inner) ? span : "[text]"
-      )
-      // a single quote is an apostrophe far more often than a quote mark, so
-      // this one only counts as a span when it is delimited on both sides -
-      // otherwise "you're not a bot" and "don't" would pair up across a
-      // perfectly safe sentence
-      .replace(/(^|[\s(])'([^']*)'(?=$|[\s).,;:!?])/g, (span, lead, inner) =>
-        MARKERS_ONLY.test(inner) ? span : `${lead}[text]`
-      )
-  )
-}
-
-/**
- * make a free-text message safe to send, or refuse to send it
- * @param {string} value - the caller's text, already through redactLogLine
- * @returns {{text: string, refused: boolean}}
- */
-function scrubText(value) {
-  let text = value
-
-  for (const [pattern, replacement] of TEXT_SCRUBS) {
-    text = text.replace(pattern, replacement)
-  }
-
-  text = scrubQuotedSpans(text)
-
-  return TEXT_UNSAFE.some((pattern) => pattern.test(text))
-    ? { text: TEXT_PLACEHOLDER, refused: true }
-    : { text, refused: false }
-}
-
-/**
- * check a value against its kind, returning the value to send.
- *
- * never coerces and never stringifies: the input may be a hostile object from
- * the renderer, and asking it for a string is asking it to run code.
- * @param {string} kind
- * @param {*} value
- * @param {string} key - the property name, for the per-property vocabularies
- * @returns {{ok: boolean, value?: *, normalized?: boolean, because?: string}}
- */
-function checkKind(kind, value, key) {
-  switch (kind) {
-    case "bool":
-      return typeof value === "boolean" ? { ok: true, value } : { ok: false }
-
-    case "number":
-      // isFinite does not coerce, so "5" and NaN both fail here
-      return Number.isFinite(value) && value >= 0 && value <= MAX_NUMBER
-        ? { ok: true, value }
-        : { ok: false }
-
-    /**
-     * a whole number of playlist items, 0 to the cap.
-     *
-     * isInteger rather than isFinite, because half a video was not saved: a
-     * fraction here is a caller that computed something rather than counted it,
-     * and the same is true of a count above the cap - a run covers at most
-     * PLAYLIST_MAX_ITEMS videos, so anything larger did not come from one.
-     */
-    case "count":
-      return Number.isInteger(value) &&
-        value >= 0 &&
-        value <= PLAYLIST_MAX_ITEMS
-        ? { ok: true, value }
-        : { ok: false }
-
-    // a finite set of values we own. an empty one is reported differently,
-    // because "nobody has declared these yet" needs a different fix from
-    // "that is not one of them"
-    case "vocabulary": {
-      const vocabulary = VOCABULARY_BY_PROPERTY.get(key)
-
-      if (!vocabulary) return { ok: false }
-      if (vocabulary.size === 0) return { ok: false, because: "empty" }
-      if (typeof value !== "string") return { ok: false }
-
-      return vocabulary.has(value) ? { ok: true, value } : { ok: false }
-    }
-
-    // heights and bitrates are matched because they grow; everything else
-    // extractQuality can return is listed because it does not
-    case "quality":
-      if (typeof value !== "string") return { ok: false }
-
-      return QUALITY_HEIGHT_PATTERN.test(value) ||
-        QUALITY_BITRATE_PATTERN.test(value) ||
-        QUALITY_VALUES.has(value)
-        ? { ok: true, value }
-        : { ok: false }
-
-    // the one kind that rewrites rather than rejects. a name we do not know
-    // is far more likely to be an unsupported site than a leak, and which
-    // sites people try is data worth having - so the event goes out carrying
-    // "unsupported" rather than being dropped over its platform. a non-string
-    // is still a plain drop: that is a caller bug, not an unknown site.
-    case "platform":
-      if (typeof value !== "string") return { ok: false }
-      if (KNOWN_PLATFORMS.has(value)) return { ok: true, value }
-      return { ok: true, value: PLATFORM_UNSUPPORTED, normalized: true }
-
-    case "version":
-      return typeof value === "string" && VERSION_PATTERN.test(value)
-        ? { ok: true, value }
-        : { ok: false }
-
-    case "bucket":
-      return typeof value === "string" && BUCKET_PATTERN.test(value)
-        ? { ok: true, value }
-        : { ok: false }
-
-    case "locale":
-      if (typeof value !== "string") return { ok: false }
-
-      return value === LOCALE_UNKNOWN || LOCALE_PATTERN.test(value)
-        ? { ok: true, value }
-        : { ok: false }
-
-
-    // scrubbed, then re-checked, and only then truncated: cutting first would
-    // leave the front half of a path behind, which no pattern then matches
-    case "text": {
-      if (typeof value !== "string") return { ok: false }
-
-      const scrubbed = scrubText(redactLogLine(value))
-
-      return {
-        ok: true,
-        value: scrubbed.text.slice(0, MAX_TEXT_LENGTH),
-        // a refusal is reported the way a normalization is, naming only the
-        // placeholder - which is a reserved constant - and never the text we
-        // could not vouch for
-        normalized: scrubbed.refused
-      }
-    }
-
-    default:
-      return { ok: false }
-  }
-}
-
-/**
- * a printable label for something that may not be printable. a Symbol event
- * name or an object with a throwing toString would otherwise blow up the very
- * console.warn meant to report it - out of telemetry and into a download.
- */
-function safeLabel(value) {
-  try {
-    return String(value)
-  } catch {
-    return "<unprintable>"
-  }
-}
+  ALLOWED_BY_EVENT,
+  ALLOWED_PROPERTIES,
+  ENVIRONMENT_DEVELOPMENT,
+  ENVIRONMENT_PRODUCTION,
+  KIND_BY_PROPERTY,
+  PROPERTY_KINDS
+} = require("./analytics/schema")
+const { checkKind, safeLabel } = require("./analytics/validate")
+const {
+  checkSuperProperties,
+  defaultCreateClient,
+  drainClient,
+  isPackagedBuild,
+  kernelVersion,
+  readLocale,
+  retireClient
+} = require("./analytics/environment")
 
 // describeError moved to utils/analytics-helpers.js: the ipc layer and the
 // download runner make the same never-throw promise this module does, and were
 // each guarding it with a weaker spelling of the same thing
-
-// electron is not present under jest, so the locale lookup must degrade
-// instead of exploding at require time
-function readLocale() {
-  try {
-    const { app } = require("electron")
-    return typeof app?.getLocale === "function"
-      ? app.getLocale()
-      : LOCALE_UNKNOWN
-  } catch {
-    return LOCALE_UNKNOWN
-  }
-}
-
-/**
- * whether this is a build somebody installed, rather than one run from source.
- *
- * this is the production indicator, and it used to be NODE_ENV === "production"
- * - which nothing in this repo ever sets. the dev scripts set "development",
- * electron-builder puts no NODE_ENV into a packaged app's environment at all,
- * and there is no .env, so every build a user actually installed evaluated that
- * gate to false and sent nothing. app.isPackaged is electron's own answer to
- * the same question and is true in exactly the builds this is about.
- *
- * NODE_ENV is deliberately not kept alongside it. it is not merely redundant:
- * it is set by tooling for reasons of its own, so keeping it as an alternative
- * signal means a machine that happens to carry NODE_ENV=production - a build
- * step, a CI shell, an exported variable - starts reporting into the production
- * project from source. that is the same class of mistake inverted, and one
- * signal that answers the question beats two where the spare one can only be
- * wrong.
- *
- * required lazily, the way readLocale() does and for the same reason: electron
- * is not present under jest, and a require at module top would take every test
- * in this file down with it.
- *
- * it fails CLOSED - unreadable means not packaged, which means no telemetry.
- * that cannot cost a real packaged build anything, because in one the electron
- * module is part of the runtime and this lookup cannot fail; the only place it
- * can fail is outside electron, where sending would be wrong anyway. failing
- * open would invert exactly that: a script, a test or a tool that constructed
- * this class would be treated as a shipped install.
- */
-function isPackagedBuild() {
-  try {
-    const { app } = require("electron")
-    return app?.isPackaged === true
-  } catch {
-    return false
-  }
-}
-
-/**
- * the kernel release, cut back to the part of it that is a version.
- *
- * os.release() is already one on darwin ("25.5.0") and on windows
- * ("10.0.22631"), but on linux it carries the distribution's packaging behind
- * a dash - "5.15.0-91-generic", "6.6.9-200.fc39.x86_64". no grammar admits
- * that tail without also admitting "1-holiday.mp4", which is the
- * counterexample VERSION_PATTERN was narrowed against in the first place; and
- * validating os.release() as-is would simply drop os_version on every linux
- * install, silently, for the life of the property.
- *
- * so it is cut rather than matched. the kernel version is the answer to "which
- * os version"; the suffix is packaging that differs per distribution and makes
- * the two harder to compare rather than easier.
- *
- * @returns {string|null} the leading numeric release, or null if there is none
- */
-function kernelVersion() {
-  const match = /^\d+(?:\.\d+)*/.exec(os.release())
-  return match ? match[0] : null
-}
-
-/**
- * check the module's own super properties the way a caller's bag is checked.
- *
- * they were never checked before, and their placement is exactly the argument
- * for checking them: capture() spreads them last, over the validated bag, onto
- * every event the app sends. a leak in one is the most amplified leak
- * available here, and nothing in ALLOWED_PROPERTIES governs them - that list
- * says what a caller may send, and no caller sends these.
- *
- * a value that fails its kind is dropped, exactly as a caller's would be. that
- * costs one dimension on the events that follow and nothing else: distinctId
- * is a separate field on the message rather than a super property, so the
- * events stay attributable, and every other property is unaffected.
- *
- * @param {Object} candidates - the property bag init() assembled
- * @returns {Object} the ones that passed
- */
-function checkSuperProperties(candidates) {
-  const safe = {}
-
-  for (const [key, value] of Object.entries(candidates)) {
-    const kind = KIND_BY_PROPERTY.get(key)
-    const checked = checkKind(kind, value, key)
-
-    if (!checked.ok) {
-      // the key and the kind, never the value - the same rule capture() keeps,
-      // for the same reason: the value is the thing we could not vouch for
-      console.warn(
-        `analytics: dropped the ${key} super property, expected ${kind}`
-      )
-      continue
-    }
-
-    safe[key] = checked.value
-  }
-
-  return safe
-}
-
-/**
- * best-effort drain of one client
- *
- * takes the client rather than reading `this.client`, because the opt-out path
- * detaches it before draining: the flush has to run against a client the
- * instance no longer holds, and a version reading the field would find null
- * there and quietly do nothing.
- *
- * @param {Object|null} client - a posthog client, or nothing
- */
-async function drainClient(client) {
-  if (!client) return
-
-  try {
-    await client.flush()
-  } catch (error) {
-    console.warn("analytics flush failed:", describeError(error))
-  }
-}
-
-/**
- * retire a client the instance has already let go of
- *
- * shutdown() rather than flush(), for three reasons - none of which is that
- * flush() leaks a timer. it does not: flush() reaches _flush(), which calls
- * clearFlushTimer() first thing (@posthog/core, posthog-core-stateless.js).
- * what shutdown() adds over it is (1) it drains in a LOOP until the queues are
- * empty rather than making one pass, (2) it waits on the in-flight sends
- * before it starts, and (3) it bounds the whole thing with the timeout it is
- * handed - which is the only cap left that means anything, now that nobody
- * waits on the result, and it is a cap we get without arming a timer of our
- * own. its documented caveat, do not reuse an instance after shutting it down,
- * is exactly this contract: the client is detached, and an opt-in builds a
- * fresh one through init().
- *
- * the flush() fallback is for a client injected by a test that has no
- * shutdown, so the drain still happens rather than being reported as broken.
- *
- * @param {Object|null} client - the client the instance no longer holds
- */
-async function retireClient(client) {
-  if (!client) return
-
-  if (typeof client.shutdown !== "function") {
-    await drainClient(client)
-    return
-  }
-
-  try {
-    await client.shutdown(OPT_OUT_FLUSH_TIMEOUT_MS)
-  } catch (error) {
-    console.warn("analytics drain failed:", describeError(error))
-  }
-}
-
-function defaultCreateClient(key, host) {
-  const { PostHog } = require("posthog-node")
-  return new PostHog(key, {
-    host,
-    // do not delete this as redundant - the sdk's own jsdoc says it defaults
-    // to false, and that is wrong. the compiled source reads
-    // `options.disableGeoip ?? true` (@posthog/core, posthog-core-stateless),
-    // confirmed by constructing a client both ways: omitted gives true.
-    // the node sdk assumes a server, where geoip on the server's own ip is
-    // meaningless. in electron the machine IS the client, so without this
-    // line country/region/city silently never arrive.
-    disableGeoip: false,
-    // the same trap one option over, and the same reason. the node sdk assumes
-    // a server, so `isServer ?? true` (posthog-node, client.js) attaches
-    // `$is_server: true` to every event unless this says otherwise. in electron
-    // the machine IS the client, so the default is not merely undisclosed - it
-    // is wrong, and it would label every event in the project as server-side.
-    //
-    // note this omits the property rather than sending `false`:
-    // getCommonEventProperties() only sets it when the option is truthy
-    isServer: false,
-    flushAt: 20,
-    flushInterval: 10000
-  })
-}
 
 class Analytics {
   /**
@@ -1289,16 +101,19 @@ class Analytics {
       // read once, not per event: the store deliberately re-reads settings on
       // every call so the settings ui cannot go stale against it
       this.installId = await this.settingsStore.getInstallId()
-      this.superProperties = checkSuperProperties({
-        app_version: getAppVersion(),
-        os: process.platform,
-        os_version: kernelVersion(),
-        arch: process.arch,
-        locale: readLocale(),
-        environment: this.isProductionBuild
-          ? ENVIRONMENT_PRODUCTION
-          : ENVIRONMENT_DEVELOPMENT
-      })
+      this.superProperties = checkSuperProperties(
+        {
+          app_version: getAppVersion(),
+          os: process.platform,
+          os_version: kernelVersion(),
+          arch: process.arch,
+          locale: readLocale(),
+          environment: this.isProductionBuild
+            ? ENVIRONMENT_PRODUCTION
+            : ENVIRONMENT_DEVELOPMENT
+        },
+        redactLogLine
+      )
 
       // the engine is probed once per run, so whatever it reported has to be
       // put back after the rebuild - nothing would set it a second time
@@ -1340,7 +155,12 @@ class Analytics {
   setEngineVersion(version) {
     if (!version) return
 
-    const checked = checkKind("version", version, "engine_version")
+    const checked = checkKind(
+      "version",
+      version,
+      "engine_version",
+      redactLogLine
+    )
 
     if (!checked.ok) {
       console.warn("analytics: ignored an engine version, expected version")
@@ -1370,10 +190,13 @@ class Analytics {
    * @param {Object} environment - {denoPresent, potProvider}
    */
   setPotEnvironment({ denoPresent, potProvider } = {}) {
-    const checked = checkSuperProperties({
-      deno_present: Boolean(denoPresent),
-      pot_provider: Boolean(potProvider)
-    })
+    const checked = checkSuperProperties(
+      {
+        deno_present: Boolean(denoPresent),
+        pot_provider: Boolean(potProvider)
+      },
+      redactLogLine
+    )
 
     this.potEnvironment = checked
     Object.assign(this.superProperties, checked)
@@ -1431,14 +254,14 @@ class Analytics {
           continue
         }
 
-        const checked = checkKind(kind, value, key)
+        const checked = checkKind(kind, value, key, redactLogLine)
 
         if (!checked.ok) {
           // the key and the expected kind, never the value: the value is the
           // suspected pii, and this warning may end up in a log a user sends us
           console.warn(
             checked.because === "empty"
-              ? `analytics: dropped ${safeLabel(key)} on ${label} - its vocabulary is empty. add the value to PROPERTY_VOCABULARIES.${safeLabel(key)} in services/analytics.js before sending it.`
+              ? `analytics: dropped ${safeLabel(key)} on ${label} - its vocabulary is empty. add the value to PROPERTY_VOCABULARIES.${safeLabel(key)} in services/analytics/schema.js before sending it.`
               : `analytics: dropped ${safeLabel(key)} on ${label}, expected ${kind}`
           )
           continue
